requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==23.1.2
Automat==24.8.1
blinker==1.8.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.5
daphne==4.1.2
exceptiongroup==1.3.0
Flask==3.0.3
gunicorn==23.0.0
h11==0.16.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==2.1.5
packaging==25.0
priority==2.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyMySQL==1.1.1
pyOpenSSL==25.1.0
python-dotenv==1.0.1
requests==2.32.3
service-identity==24.2.0
taskgroup==0.2.2
tomli==2.2.1
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
Werkzeug==3.0.6
wsproto==1.2.0
zipp==3.20.2
zope.interface==7.2
